+++ app/src/routes/api/v1/subscription.router.js
@@ -353,7 +353,7 @@ const isAdmin = async (ctx, next) => {
     const loggedUser = SubscriptionsRouter.getUser(ctx);
 
     if (!loggedUser || USER_ROLES.indexOf(loggedUser.role) === -1) {
-        ctx.throw(401, 'Not authorized');
+        ctx.throw(401, 'Unauthorized');
         return;
     }
     if (loggedUser.role !== 'ADMIN') {
@@ -403,7 +403,7 @@ const hasValidLoggedUser = (ctx) => {
 
 const validateLoggedUserAuth = async (ctx, next) => {
     if (!hasLoggedUser(ctx)) {
-        ctx.throw(401, 'Not authorized');
+        ctx.throw(401, 'Unauthorized');
         return;
     }
 
@@ -417,7 +417,7 @@ const validateLoggedUserAuth = async (ctx, next) => {
 
 const validateMicroserviceAuth = async (ctx, next) => {
     if (!isMicroservice(ctx)) {
-        ctx.throw(401, 'Not authorized');
+        ctx.throw(401, 'Unauthorized');
         return;
     }
 
@@ -426,11 +426,21 @@ const validateMicroserviceAuth = async (ctx, next) => {
 
 const validateLoggedUserOrMicroserviceAuth = async (ctx, next) => {
     if (!(hasLoggedUser(ctx) || hasValidLoggedUser(ctx)) && !isMicroservice(ctx)) {
-        ctx.throw(401, 'Not authorized');
+        ctx.throw(401, 'Unauthorized');
         return;
     }
 
     await next();
+};
+
+const isAdminOrMicroservice = async (ctx, next) => {
+    const loggedUser = SubscriptionsRouter.getUser(ctx);
+
+    if (isMicroservice(ctx) || (loggedUser && loggedUser.role && ['ADMIN', 'SUPERADMIN', 'MICROSERVICE'].includes(loggedUser.role))) {
+        return next();
+    }
+
+    return ctx.throw(401, 'Unauthorized');
 };
 
 router.post('/', SubscriptionsRouter.createSubscription);
@@ -449,7 +459,7 @@ router.delete('/:id', validateLoggedUserOrMicroserviceAuth, subscriptionExists(t
 router.post('/notify-updates/:dataset', SubscriptionsRouter.notifyUpdates);
 router.post('/test-email-alerts', isAdmin, SubscriptionsRouter.testEmailAlert);
 router.post('/check-hook', SubscriptionsRouter.checkHook);
-router.get('/user/:userId', validateMicroserviceAuth, SubscriptionsRouter.findUserSubscriptions);
+router.get('/user/:userId', isAdminOrMicroservice, SubscriptionsRouter.findUserSubscriptions);
 router.post('/find-by-ids', validateMicroserviceAuth, SubscriptionsRouter.findByIds);
 
 module.exports = router;